Rename 'TypedProgram.requirement(referredBy:)' to 'traitMember(referredBy:)'
